requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
alembic==1.14.1
async-timeout==5.0.1
attrs==24.3.0
Authlib==1.4.0
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cs50==9.4.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Redmail==0.3.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
kivy==2.2.1
kivymd==1.1.1
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.19.0
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
redmail==0.6.0
requests==2.32.3
SQLAlchemy==2.0.36
sqlparse==0.5.3
termcolor==2.5.0
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0
